requirements.txt
pandas
langchain
langchain-core
langchain-ollama
langchain-chroma
langchain-community
ollama
unstructured
unstructured[pdf,ppt,doc,image]
firebase-admin
pyrebase4